+++ xscen/extract.py
@@ -131,7 +131,7 @@ def extract_dataset(
     xr_open_kwargs: dict = None,
     xr_combine_kwargs: dict = None,
     preprocess: Callable = None,
-    resample_method: Optional[str] = None,
+    resample_methods: Optional[dict] = None,
 ) -> Union[dict, xr.Dataset]:
     """
     Takes one element of the output of `search_data_catalogs` and returns a dataset,
@@ -163,8 +163,10 @@ def extract_dataset(
       will be passed to `xr.combine_by_coords`.
     preprocess : callable, optional
       If provided, call this function on each dataset prior to aggregation.
-    resample_method : {'mean', 'min', 'max', 'sum', 'wind_direction'}, optional
-      The resampling method. If None (default), it is guessed from the variable name and frequency,
+    resample_methods : dict, optional
+      Dictionary where the keys are the variables and the values are the resampling method.
+      Options for the resampling method are {'mean', 'min', 'max', 'sum', 'wind_direction'}.
+      If the method is not given for a variable, it is guessed from the variable name and frequency,
       using the mapping in CVs/resampling_methods.json. If the variable is not found there,
       "mean" is used by default.
 
@@ -187,6 +189,8 @@ def extract_dataset(
         buffer: float, optional
             Multiplier to apply to the model resolution.
     """
+    resample_methods = resample_methods or {}
+
     # Checks
     # must have all the same processing level and same id
     unique = catalog.unique()
@@ -291,7 +295,7 @@ def extract_dataset(
                                     da,
                                     variables_and_freqs[var_name],
                                     ds=ds_ts,
-                                    method=resample_method,
+                                    method=resample_methods.get(var_name, None),
                                 )
                             }
                         )